+++ src/helpers/ExtensionSettings.ts
@@ -1,0 +1,6 @@
+export const CONFIG_KEY = "spfxLocalization";
+export const CONFIG_CSV_FILELOCATION = "csvFileLocation";
+export const CONFIG_CSV_DELIMITER = "csvDelimiter";
+export const CONFIG_AUTO_EXPORT = "autoCsvExport";
+
+export const OPTION_IMPORT_ALL = "Import to all";