+++ src/helpers/ProjectFileHelper.ts
@@ -1,8 +1,9 @@
 import * as vscode from 'vscode';
-import { Config } from './../commands/Config';
+import * as path from 'path';
+import { Config, LocalizedResourceValue } from '../models/Config';
 
 export default class ProjectFileHelper {
-  
+    
   /**
    * Fetch the project config file
    */
@@ -43,5 +44,32 @@ export default class ProjectFileHelper {
     }
 
     return null;
+  }
+
+
+  /**
+   * Retrieve the resource path for the file
+   * 
+   * @param resx 
+   */
+  public static getResourcePath(resx: LocalizedResourceValue): string {
+    // Create the key in the localized resource file
+    let resourcePath = resx.value.substring(0, resx.value.lastIndexOf('/'));
+    // Check if the path starts with 'lib/', if this is the case, it needs to be changed to 'src/'
+    if (resourcePath.startsWith("lib/")) {
+      resourcePath = resourcePath.replace("lib/", "src/");
+    }
+    return resourcePath;
+  }
+
+  /**
+   * Get the absolute path for the file
+   * 
+   * @param fileLocation 
+   */
+  public static getAbsPath(fileLocation: string): any {
+    // Create the file path
+    const rootPath = vscode.workspace.rootPath || __dirname;
+    return path.join(rootPath, fileLocation);
   }
 }