+++ src/helpers/TextHelper.ts
@@ -7,6 +7,11 @@ export default class TextHelper {
    * @param value 
    */
   public static stripQuotes(value: string): string {
+    // Strip the comma
+    if (value.endsWith(",")) {
+      value = value.substring(0, value.length - 1);
+    }
+
     if ((value.startsWith(`'`) && value.endsWith(`'`)) || 
         (value.startsWith(`"`) && value.endsWith(`"`)) || 
         (value.startsWith("`") && value.endsWith("`"))) {