+++ src/models/Config.ts
@@ -25,3 +25,8 @@ export interface Component {
   entrypoint: string;
   manifest: string;
 }
+
+export interface LocalizedResourceValue {
+  key: string;
+  value: string;
+}