+++ src/models/LocaleKeyValue.ts
@@ -1,0 +1,4 @@
+export interface LocaleKeyValue {
+  key: string;
+  value: string;
+}